Rename configuration package to config
