+++ config/src/main/java/com/dumbdogdiner/stickyapi/common/config/Configuration.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2020-2021 DumbDogDiner <dumbdogdiner.com>. All rights reserved.
  * Licensed under the MIT license, see LICENSE for more information...
  */
-package com.dumbdogdiner.stickyapi.common.configuration;
+package com.dumbdogdiner.stickyapi.common.config;
 
 /**
  * Base interface for a configuration.
